+++ model/tangent_prop.py
@@ -79,7 +79,7 @@ class TangentPropClassifier(BaseClassifierModel, BaseNeuralNet):
         return y_pred
 
     def predict_proba(self, X):
-        X = X.astype(np.float32)
+        # X = X.astype(np.float32)
         with torch.no_grad():
             X_torch = to_torch(X, cuda=self.cuda_flag)
             logits = self.net(X_torch)
